+++ lib/avrolution/railtie.rb
@@ -1,0 +1,14 @@
+module Avrolution
+  class Railtie < Rails::Railtie
+
+    initializer 'avrolution.configure' do
+      Avrolution.configure do |config|
+        config.root = Rails.root
+      end
+    end
+
+    rake_tasks do
+      load File.expand_path('../rake/rails_avrolution.rake', __FILE__)
+    end
+  end
+end