+++ spec/avrolution_spec.rb
@@ -1,7 +1,0 @@
-require 'spec_helper'
-
-describe Avrolution do
-  it "has a version number" do
-    expect(Avrolution::VERSION).not_to be nil
-  end
-end