+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 # avrolution
 
 ## v0.1.0
-- Initial version
+- Add rake task to check the compatibility of schemas against a schema registry.